Another test...

+++ gpytorch/kernels/matern_kernel.py
@@ -107,4 +107,4 @@ class MaternKernel(Kernel):
                 constant_component = (math.sqrt(5) * distance).add(1).add(5.0 / 3.0 * distance ** 2)
             return constant_component * exp_component
         return MaternCovariance().apply(x1, x2, self.lengthscale, self.nu,
-                                        lambda x1, x2: self._covar_dist(x1, x2, **params))
+                                        lambda x1, x2: self._covar_dist(x1.half().float(), x2.half().float(), **params).half().float())
